Internal functions for coordinate transformations are now private API (#7113)
