+++ sunpy/coordinates/__init__.py
@@ -19,10 +19,10 @@ below (see `astropy.coordinates.builtin_frames`).
 """
 
 from . import sun
+from ._transformations import _make_sunpy_graph, propagate_with_solar_surface, transform_with_sun_center
 from .ephemeris import *
 from .frames import *
 from .metaframes import *
-from .transformations import _make_sunpy_graph, propagate_with_solar_surface, transform_with_sun_center
 from .wcs_utils import *
 
 __doc__ += _make_sunpy_graph()
