+++ sunpy/physics/differential_rotation.py
@@ -6,8 +6,13 @@ import astropy.units as u
 from astropy.coordinates import BaseCoordinateFrame, Longitude, SkyCoord
 from astropy.time import TimeDelta
 
-from sunpy.coordinates import Heliocentric, HeliographicStonyhurst, Helioprojective, get_earth
-from sunpy.coordinates.transformations import transform_with_sun_center
+from sunpy.coordinates import (
+    Heliocentric,
+    HeliographicStonyhurst,
+    Helioprojective,
+    get_earth,
+    transform_with_sun_center,
+)
 from sunpy.map import (
     contains_full_disk,
     coordinate_is_on_solar_disk,
